third_party/blink/web_tests/platform/mac/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-flexbox-expected.txt

This is a testharness.js-based test.
FAIL Flex assert_equals: height expected "18px" but got "162px"
FAIL Inline flex assert_equals: height expected "18px" but got "162px"
FAIL Dynamic change of flex-direction assert_equals: expected 44 but got 26
Harness: the test ran to completion.
